Добавлена фильтрация на главной странице по специализациям, стажу и технологиям проектов

+++ src/main/java/org/example/springappportfolio/dto/ProjectSummary.java
@@ -1,7 +1,7 @@
 package org.example.springappportfolio.dto;
 
 public record ProjectSummary (
-   Integer projectId,
+   Long projectId,
    String title,
    String shortDescription,
    byte[] preview,
